feat: build uses java ffi
