+++ app/gui2/shared/ast/ffi.ts
@@ -1,13 +1,33 @@
 import type { IDataType } from 'hash-wasm/dist/lib/util'
+import { createXXHash128 } from 'hash-wasm'
+import init, { is_ident_or_operator, parse, parse_doc_to_json } from '../../rust-ffi/pkg/rust_ffi'
+import { assertDefined } from '../util/assert'
+import { isNode } from '../util/detect'
 
-declare global {
-    function parse_tree(code: string): Uint8Array;
-    function parse_doc_to_json(docs: string): string;
-    function is_ident_or_operator(code: string): number;
-    function xxHash128(input: IDataType): string;
+let xxHasher128: Awaited<ReturnType<typeof createXXHash128>> | undefined
+export function xxHash128(input: IDataType) {
+  assertDefined(xxHasher128, 'Module should have been loaded with `initializeFFI`.')
+  xxHasher128.init()
+  xxHasher128.update(input)
+  return xxHasher128.digest()
 }
 
-export async function initializeFFI(path?: string | undefined) {}
+export async function initializeFFI(path?: string | undefined) {
+  if (isNode) {
+    const fs = await import('node:fs/promises')
+    const { fileURLToPath, URL: nodeURL } = await import('node:url')
+    const buffer = fs.readFile(
+      path ?? fileURLToPath(new nodeURL('../../rust-ffi/pkg/rust_ffi_bg.wasm', import.meta.url)),
+    )
+    await init(buffer)
+  } else {
+    await init()
+  }
+  xxHasher128 = await createXXHash128()
+}
+
+// TODO[ao]: We cannot to that, because the ffi is used by cjs modules.
+// await initializeFFI()
 
 /* eslint-disable-next-line camelcase */
-export const { is_ident_or_operator, parse_doc_to_json, parse_tree, xxHash128 } = globalThis;
+export { is_ident_or_operator, parse_doc_to_json, parse as parse_tree }
